+++ src/rclcpp_spk/src/rclcpp/parameter_map.cpp
@@ -35,7 +35,7 @@ rclcpp::parameter_map_from(const rcl_params_t * const c_params)
 
   // Convert c structs into a list of parameters to set
   ParameterMap parameters;
-  for (size_t n = 0; n < c_params->num_nodes; ++n) {
+  for (int n = 0; n < c_params->num_nodes; ++n) {
     const char * c_node_name = c_params->node_names[n];
     if (NULL == c_node_name) {
       throw InvalidParametersException("Node name at index " + std::to_string(n) + " is NULL");
@@ -54,7 +54,7 @@ rclcpp::parameter_map_from(const rcl_params_t * const c_params)
     std::vector<Parameter> & params_node = parameters[node_name];
     params_node.reserve(c_params_node->num_params);
 
-    for (size_t p = 0; p < c_params_node->num_params; ++p) {
+    for (int p = 0; p < c_params_node->num_params; ++p) {
       const char * const c_param_name = c_params_node->parameter_names[p];
       if (NULL == c_param_name) {
         std::string message(
@@ -86,7 +86,7 @@ rclcpp::parameter_value_from(const rcl_variant_t * const c_param_value)
     const rcl_byte_array_t * const byte_array = c_param_value->byte_array_value;
     std::vector<uint8_t> bytes;
     bytes.reserve(byte_array->size);
-    for (size_t v = 0; v < byte_array->size; ++v) {
+    for (int v = 0; v < byte_array->size; ++v) {
       bytes.push_back(byte_array->values[v]);
     }
     return ParameterValue(bytes);
@@ -94,7 +94,7 @@ rclcpp::parameter_value_from(const rcl_variant_t * const c_param_value)
     const rcl_bool_array_t * const bool_array = c_param_value->bool_array_value;
     std::vector<bool> bools;
     bools.reserve(bool_array->size);
-    for (size_t v = 0; v < bool_array->size; ++v) {
+    for (int v = 0; v < bool_array->size; ++v) {
       bools.push_back(bool_array->values[v]);
     }
     return ParameterValue(bools);
@@ -102,7 +102,7 @@ rclcpp::parameter_value_from(const rcl_variant_t * const c_param_value)
     const rcl_int64_array_t * const int_array = c_param_value->integer_array_value;
     std::vector<int64_t> integers;
     integers.reserve(int_array->size);
-    for (size_t v = 0; v < int_array->size; ++v) {
+    for (int v = 0; v < int_array->size; ++v) {
       integers.push_back(int_array->values[v]);
     }
     return ParameterValue(integers);
@@ -110,7 +110,7 @@ rclcpp::parameter_value_from(const rcl_variant_t * const c_param_value)
     const rcl_double_array_t * const double_array = c_param_value->double_array_value;
     std::vector<double> doubles;
     doubles.reserve(double_array->size);
-    for (size_t v = 0; v < double_array->size; ++v) {
+    for (int v = 0; v < double_array->size; ++v) {
       doubles.push_back(double_array->values[v]);
     }
     return ParameterValue(doubles);
@@ -118,7 +118,7 @@ rclcpp::parameter_value_from(const rcl_variant_t * const c_param_value)
     const rcutils_string_array_t * const string_array = c_param_value->string_array_value;
     std::vector<std::string> strings;
     strings.reserve(string_array->size);
-    for (size_t v = 0; v < string_array->size; ++v) {
+    for (int v = 0; v < string_array->size; ++v) {
       strings.emplace_back(string_array->data[v]);
     }
     return ParameterValue(strings);
